Fix bin/gitlab_before.sh

See merge request puppet/psick!170

+++ bin/gitlab_before.sh
@@ -5,16 +5,18 @@ script_dir="$(dirname $0)"
 . "${script_dir}/functions"
 
 # r10k config file 
-configfile=${1:-bin/config/r10k.yaml}
+# configfile=${1:-bin/config/r10k.yaml}
 # Location of keys to copy into the local repository (removed from gilab_after.sh
 eyamlkeyloc=${1:-/etc/eyaml/keys}
 
-echo "Setup keys"
-mkdir -p ${repo_dir}/keys
-cp -f ${eyamlkeyloc}/* ${repo_dir}/keys/
+if [ -f "$eyamlkeyloc" ]; then
+  echo "Setup keys"
+  mkdir -p ${repo_dir}/keys
+  cp -f ${eyamlkeyloc}/* ${repo_dir}/keys/
+fi
 
 if [ -f "$configfile" ]; then
-	config="-c ${configfile}"
+  config="-c ${configfile}"
 fi
 echo 
 cd $repo_dir
